requirements.txt
streamlit
langchain
langchain-openai
langchain-cli
openai
streamlit-chat
python-dotenv